Add feature: hr manager can create job offer public state.
This will serve as a contract to keep track of nbr of applications.

So upon accepting/rejecting a job application, this state will change,
providing users with valuable observations, like current applciations in progress.

NOTE: we ensure we create a job offer public state before creating a job offer,
but there is no code which ensure this relationship.

We are relying on job offer and public state to have same key, thus giving us the
possibility to excercise choices from one to the other.